requirements.txt
aiohappyeyeballs>=2.4.3
aiohttp>=3.10.10
aiosignal>=1.3.1
altair>=5.4.1
annotated-types>=0.7.0
anyio>=4.6.2.post1
attrs>=24.2.0
beautifulsoup4>=4.12.3
bcrypt>=4.2.0
blinker>=1.8.2
cachetools>=5.5.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
click>=8.1.7
colorama>=0.4.6
dataclasses-json>=0.6.7
distro>=1.9.0
Flask>=3.0.3
Flask-Bcrypt>=1.0.1
flask-cors>=5.0.0
Flask-JWT-Extended>=4.6.0
Flask-SQLAlchemy>=3.1.1
frozenlist>=1.5.0
gitdb>=4.0.11
GitPython>=3.1.43
google-ai-generativelanguage>=0.6.10
google-api-core>=2.20.0
google-api-python-client>=2.154.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.8.3
google-search-results>=2.4.2
serpapi>=0.1.5
googleapis-common-protos>=1.65.0
greenlet>=3.1.1
grpcio>=1.67.1
grpcio-status>=1.67.1
h11>=0.14.0
httpcore>=1.0.6
httplib2>=0.22.0
httpx>=0.27.2
httpx-sse>=0.4.0
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.4
jiter>=0.7.0
jsonpatch>=1.33
jsonpointer>=3.0.0
jsonschema>=4.23.0
jsonschema-specifications>=2023.12.1
langchain>=0.3.3
langchain-community>=0.3.2
langchain-core>=0.3.10
langchain-google-genai>=1.0.8
langchain-text-splitters>=0.3.0
# LangSmith and LangGraph packages removed per cleanup
MarkupSafe>=3.0.1
marshmallow>=3.23.0
multidict>=6.1.0
mypy_extensions>=1.0.0
narwhals>=1.9.3
numpy>=2.1.1
openai>=1.52.2
orjson>=3.10.7
packaging>=24.1
pandas>=2.2.3
pdfplumber>=0.11.4
pillow>=10.4.0
propcache>=0.2.0
proto-plus>=1.25.0
protobuf>=5.28.2
pyarrow>=17.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pydantic>=2.9.2
pydantic-settings>=2.6.0
pydantic_core>=2.23.4
pydeck>=0.9.1
PyJWT>=2.9.0
pyparsing>=3.1.4
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.2
PyYAML>=6.0.2
referencing>=0.35.1
regex>=2024.9.11
requests>=2.32.3
requests-toolbelt>=1.0.0
rpds-py>=0.20.0
rsa>=4.9
six>=1.16.0
smmap>=5.0.1
sniffio>=1.3.1
SQLAlchemy>=2.0.35
streamlit>=1.39.0
tenacity>=9.0.0
tiktoken>=0.8.0
toml>=0.10.2
tornado>=6.4.1
tqdm>=4.66.5
typing-inspect>=0.9.0
typing-inspection>=0.1.0
typing_extensions>=4.12.2
tzdata>=2024.2
uritemplate>=4.1.1
PyMuPDF>=1.24.11
urllib3>=2.2.3
watchdog>=5.0.3
Werkzeug>=3.0.4
yarl>=1.13.1
zstandard>=0.23.0
langchain-openai>=0.2.3
Flask-Session>=0.8.0
gunicorn>=23.0.0
PyMySQL
redis>=5.0.0
openpyxl
pymongo>=4.6.0
gridfs